Cleaner method for DOM testing
Using jQuery CSS selectors as the method of validation lets us define what parts of the relationship are important as well as checking the expected value. I created a few helper methods to simplify the registration of tests using ``it()`` so that only the most relevant part is listed in the behavior test file. This does make it a little less readable (doesn't follow the "it should ..." syntax), but makes it much easier to write a greater volume of specific tests.

This example is probably a bit extreme, but I wanted to show how each aspect of the expected DOM structure could be tested individually. For example

```javascript
expectDom.found(getComponentElement, '> li:visible:eq(0) > button');
expectDom.found(getComponentElement, '> li:visible:eq(0) > button:eq(0).tree-item-name');
expectDom.found(getComponentElement, '> li:visible:eq(0) > button:eq(0)[type=button]');
```

could instead be

```javascript
expectDom.found(getComponentElement, '> li:visible:eq(0) > button:eq(0).tree-item-name[type=button]');
```

but I felt it made it easier to read to split up the properties being tested. In this case, the presence of the class and the ``type`` attribute.

The important part is that the use of jQuery selectors, such as ``:visible`` and ``:eq(N)`` along with the more common CSS selectors such as ``>`` and ``+`` allow us to test the structure of the DOM from the component's element downward with less difficulty, while still allowing for flexibility. For example, if the ``button`` above didn't need to be an immediate child of the ``li``, then instead of writing ``li:visible:eq(0) > button`` I could write ``li:visible:eq(0) button``.