Fixes #917 - guarantee uniq react key for datatable items
Not sure if we want to add a warning, since this is a new pattern in
DSR. Currently we do not warn the user in any other area. Instead I
would like to propose to actually use the shortid.generate method that
already existed on this line but never got called.

This will guarantee the uniqueness of the react item keys.

If we'd like to log a warning, I would like to suggest for us to have a
module through which dispatch warnings/errors that will do a
console.error