Fix tree selection with complex nodes
If a tree has a complex object for its nodes (deeply nested, etc), the
keyboard select fails to work properly. This is due to the complex
objects being compared by reference, instead of by value. To fix this,
we are now finding the node by id, since that is a required value for
every node.