Brand Band: Adds a new prop `backgroundSize`
There is a class is SLDS `slds-brand-band_cover` which changes the
background size from `contain` to `cover`. There was no such prop
in DSR to use that class. This commit adds that prop `backgroundSize`
to change the background size to `cover` if needed. The default size
is `contain` regardless.

Fixes #1897