Breadcrumb: Ability to add overflow menu
There was no option to add overflow menu in the current version
of breadcrumb. This commit aims at adding that functionality by
adding a new prop `parentIndex` which slices the trail into a
menu trail and a regular trail, which can be then used to render
overflow menu and regular breadcrumb. The prop can have a better
name though, as it might not be very obvious in the first glance.