Popover: adds hasNubbin prop to popover
Currently it is assumed that nubbin is necessary for a popover,
which is the default case but is not the necessary one. There should
exist a prop for this. `hasNubbin` prop has default value `true`
so that the default case is satified, and it gives more control to
the user to disable the nubbin if need be.