ColorPicker: Removes differences in SLDS Color Picker and DSR Color picker
SLDS Color picker doesn't have any nubbin or `close` button, because
the `close` button is present in the footer as `cancel` button. The
redundant `close` button is removed and also the nubbin is removed
as the context of the `dialog`/`popover` is clear enough that it
doesn't need the nubbin. This is to make the DSR Color Picker
coherent with the SLDS Color Picker.

Fixes #1724