AppLauncher: Fixes arrow direction in app launcher sections
The current app launcher sections were having the icons beside
heading that didn't change orientation based on section state.
They basically pointed in the same direction irrespective of the
state of the section. This commit fixes that issue and changes
orientation of the button icon based on state of the section.

Fixes #1905