AppLauncher: Changes icontext in tile to be consistent with heading
A very trivial issue with an equally trivial fix. The fix isn't
particularly necessary for the project but for the sake of
consistency of the elements and the data displayed by elements,
this commit changes the iconText in one of the examples to make
it consistent with the heading.

Fixes #1907