Fix AccordionPanel: application crashes if refs is not provided
refs is not a required prop for AccordinPanel component, but if we have not provided refs, refs will be undefined,
refs.summaryButton will not exist.This makes app crash.This error will occur only if we use AccordinPanel
individually(Not as a child of Accordin).That is becuase Accordin component will pass refs by default.

This issue has fixed using a ternary expression.If refs exist then buttonRef takes refs.summaryButton otherwise
it will have a null value.Along with this I have removed onClickSummary prop. as it is unused, throughout the component

Fixes: #2450