Fixing a11y issues in the Tabs component
Moved the tabIndex in the Tab component out away
from the <li> tag that has role="presentation" and to
a child <a> tag.

tabIndex=0 makes an element focusable where presentation
role removes it from the a11y tree. Presence of both of
these on the same element results into a conflict and
arbitrary behaviour by screen readers.