package.json: update license to be machine-parsable
This allows tools which care about the licensing of the source code here to more
easily deal with this package in validation of licenses.

The license picked here is based on the license for the source code, and the
expectation is that users will consult the README for licenses on other assets.